+++ backend/package.json
@@ -7,6 +7,7 @@
     "fix": "prettier --write . && eslint --fix --ext .js ."
   },
   "dependencies": {
+    "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
