+++ backend/src/router.js
@@ -9,7 +9,8 @@ const profilsControllers = require("./controllers/profilsControllers");
 const gamesControllers = require("./controllers/gamesControllers");
 const authControllers = require("./controllers/authControllers");
 
-const { checkUserData } = require("./services/auth");
+const { checkUserData } = require("./services/checkUserData");
+const { hashPassword } = require("./services/auth");
 
 // router.get("/items", itemControllers.browse);
 // router.get("/items/:id", itemControllers.read);
@@ -32,6 +33,6 @@ router.get("/profils/:id", profilsControllers.read);
 router.get("/games", gamesControllers.browse);
 router.get("/games/:id", gamesControllers.read);
 
-router.post("/signup", checkUserData, authControllers.signup);
+router.post("/signup", checkUserData, hashPassword, authControllers.signup);
 
 module.exports = router;
